Прога/scripts/requirements-materials.txt
# Для convert_materials.py (ноутбуки + PDF)
# + convert_bd_course_materials / convert_minor_course_materials (PDF, DOCX)
nbconvert>=7.0
nbformat>=5.0
jupyter-core
pymupdf>=1.23.0
python-docx>=1.0.0
python-pptx>=0.6.21
